requirements.txt
py-cord==1.7.3
python-dotenv==0.19.0
requests==2.26.0